Helper field type changed to byte

+++ DragonDustWeb/Models/OrderType.cs
@@ -10,6 +10,6 @@ namespace DragonDustWeb.Models
         public int Id { get; set; }
         public string Name { get; set; }
 
-        public static readonly int FreePromoCodeByEmailId = 1;
+        public static readonly byte FreePromoCodeByEmailId = 1;
     }
 }
